Création du composant externalLinks et déplacement de la logique LinkExt dans externalLinks
