+++ client/src/constants/externalLinks.ts
@@ -1,0 +1,29 @@
+interface LinkExtType {
+    instagram: string
+    facebook: string
+    youtube: string
+
+    /* Contact */
+    emailAnne: string
+    phoneAnne: string
+    phoneHebergeur: string
+
+    /* Article de loi */
+    loi1: string
+    loi2: string
+    loi3: string
+}
+
+const externalLinks: LinkExtType = {
+    instagram: "https://www.instagram.com/atelier_photo_montpellier",
+    facebook: "https://www.facebook.com",
+    youtube: "https://www.youtube.com",
+    emailAnne: "mailto:photo34000@gmail.com",
+    phoneAnne: "tel:+33652677333",
+    loi1: "https://www.legifrance.gouv.fr/loda/id/JORFTEXT000000801164#LEGIARTI000042038977",
+    loi2: "https://www.legifrance.gouv.fr/codes/article_lc/LEGIARTI000032655082",
+    loi3: "https://www.legifrance.gouv.fr/loda/id/JORFTEXT000000886460",
+    phoneHebergeur: "tel:+33444446040"
+}
+
+export default externalLinks;
